openai-streaming-client/requirements.txt
openai>=1.0.0
python-dotenv>=0.21.0
